refactor(playground): split individual plugin block
