+++ packages/agent-playground/src/components/playground/individual-plugin-block/constants.ts
@@ -1,0 +1,1 @@
+export const PERCENTAGE_MULTIPLIER = 100;
