+++ packages/agent-playground/src/components/playground/individual-plugin-block/index.tsx
@@ -1,0 +1,1 @@
+export { IndividualPluginBlock } from './individual-plugin-block';
